+++ cdc/service.go
@@ -229,13 +229,18 @@ func NewService(nodeID, dir string, clstr Cluster, cfg *Config) (*Service, error
 	}
 	srv.fifo = fifo
 
-	// Whatever is the first key in the FIFO is our initial high watermark. We assume
-	// that anything sitting in the queue has not been sent to the webhook. If that is
-	// not the case then an HWM update from other nodes in the cluster may update it
+	// Whatever is the first key in the FIFO we assume has not been sent. This ensures we meet the
+	// at least-once guarantee. So set the highwater mark to one before.
+	//
+	// In other words we assume that anything sitting in the queue has not been sent to the webhook.
+	// If that is not the case then an HWM update from other nodes in the cluster may update it
 	// (and prune the FIFO).
 	higHWM, err := fifo.FirstKey()
 	if err != nil {
 		return nil, fmt.Errorf("failed to read first key from FIFO: %w", err)
+	}
+	if higHWM > 0 {
+		higHWM -= 1
 	}
 	srv.highWatermark.Store(higHWM)
 
@@ -471,7 +476,10 @@ func (s *Service) writeToBatcher() {
 			}
 
 			s.batcher.Flush()
-			<-fc      // Wait for CDC to write to BoltDB.
+			select {
+			case <-fc: // Wait for CDC to write to BoltDB.
+			case <-s.done: // Or detect we're shutting down.
+			}
 			close(ch) // Snapshotting can proceed now.
 		case <-s.done:
 			return
@@ -504,7 +512,6 @@ func (s *Service) leaderLoop() (chan struct{}, chan struct{}) {
 
 			case ev := <-s.fifo.C:
 				if ev == nil {
-					close(done)
 					return
 				}
 				if ev.Index <= s.highWatermark.Load() {
@@ -553,7 +560,15 @@ func (s *Service) leaderLoop() (chan struct{}, chan struct{}) {
 					}
 					stats.Add(numRetries, 1)
 					s.endpointRetries.Add(1)
-					time.Sleep(retryDelay)
+
+					// Sleep, but detect any shutdown request while sleeping.
+					t := time.NewTimer(retryDelay)
+					select {
+					case <-stop:
+						t.Stop()
+						return
+					case <-t.C:
+					}
 				}
 				if sentOK {
 					s.highWatermark.Store(ev.Index)