+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [PR #2710](https://github.com/rqlite/rqlite/pull/2710): Don't run CRC checks twice on Snapshot when fast-restart is explicitly disabled.
 - [PR #2711](https://github.com/rqlite/rqlite/pull/2711): Minor fixes to the `cluster` package related to connection handling and metrics.
 - [PR #2712](https://github.com/rqlite/rqlite/pull/2712): Fixes to the `http` package including SQL rewriting control.
+- [PR #2713](https://github.com/rqlite/rqlite/pull/2713): Address potential dropped CDC event on restart.
 
 ## v10.2.6 (July 4th 2026)
 ### Implementation changes and bug fixes